+++ app/Filament/Admin/Resources/Pengembalians/Pages/CreatePengembalian.php
@@ -8,4 +8,9 @@ use Filament\Resources\Pages\CreateRecord;
 class CreatePengembalian extends CreateRecord
 {
     protected static string $resource = PengembalianResource::class;
+
+     protected function getRedirectUrl(): string
+    {
+        return $this->getResource()::getUrl('index');
+    }
 }