+++ resources/js/app.js
@@ -1,1 +1,6 @@
-//
+import Alpine from 'alpinejs';
+import './progress';
+
+window.Alpine = Alpine;
+
+Alpine.start();